gpu int8

+++ src/lex/gpu.cu
@@ -7,7 +7,7 @@ __global__ void genperm_lex_device(int n, int prefix_len, int* counter) {
   int task_idx = threadIdx.x + blockIdx.x * blockDim.x;
   int perm_s = 0;
 
-  int a[MAX_N];
+  int8_t a[MAX_N];
 
   if (idx2prefix(n, prefix_len, task_idx, a)) {
     while (true) {
@@ -47,17 +47,17 @@ class LexGpu: public PermAlgorithm<LexGpu> {
   GENERATE_CONSTRUCTOR(LexGpu)
 
  private:
-	GPU_ALGO_ARGS
+  GPU_ALGO_ARGS
 
  protected:
   void setup_() override {
-		SETUP_GPU_ALGO()
+    SETUP_GPU_ALGO()
   }
 
  public:
   template <typename F>
   void do_generate_(F&& callback)  {
-		LAUNCH_GPU_ALGO(genperm_lex_device);
+    LAUNCH_GPU_ALGO(genperm_lex_device);
   }
 };
